+++ oap-server/analyzer/meter-analyzer/src/main/java/org/apache/skywalking/oap/meter/analyzer/dsl/EntityDescription/ServiceRelationEntityDescription.java
@@ -19,14 +19,14 @@
 package org.apache.skywalking.oap.meter.analyzer.dsl.EntityDescription;
 
 import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
+import com.google.common.collect.ImmutableList;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import lombok.ToString;
 import org.apache.skywalking.oap.server.core.analysis.Layer;
 import org.apache.skywalking.oap.server.core.analysis.meter.ScopeType;
 import org.apache.skywalking.oap.server.core.source.DetectPoint;
+import org.apache.skywalking.oap.server.library.util.StringUtil;
 
 @Getter
 @RequiredArgsConstructor
@@ -38,9 +38,16 @@ public class ServiceRelationEntityDescription implements EntityDescription {
     private final DetectPoint detectPoint;
     private final Layer layer;
     private final String delimiter;
+    private final String componentIdKey;
 
     @Override
     public List<String> getLabelKeys() {
-        return Stream.concat(this.sourceServiceKeys.stream(), this.destServiceKeys.stream()).collect(Collectors.toList());
+        final ImmutableList.Builder<String> builder = ImmutableList.<String>builder()
+            .addAll(this.sourceServiceKeys)
+            .addAll(this.destServiceKeys);
+        if (StringUtil.isNotEmpty(componentIdKey)) {
+            builder.add(componentIdKey);
+        }
+        return builder.build();
     }
 }