+++ oap-server/server-alarm-plugin/src/main/java/org/apache/skywalking/oap/server/core/alarm/provider/AlarmMessageFormatter.java
@@ -20,7 +20,6 @@ package org.apache.skywalking.oap.server.core.alarm.provider;
 
 import java.util.ArrayList;
 import java.util.List;
-import org.apache.skywalking.oap.server.core.alarm.MetaInAlarm;
 
 /**
  * This is a formatter especially for alarm message.
@@ -72,17 +71,17 @@ public class AlarmMessageFormatter {
         while (match);
     }
 
-    public String format(MetaInAlarm meta) {
+    public String format(AlarmEntity alarmEntity) {
         StringBuilder message = new StringBuilder();
         for (int i = 0; i < formatSegments.size(); i++) {
             message.append(formatSegments.get(i));
             if (i != formatSegments.size() - 1) {
                 switch (valueFroms.get(i)) {
                     case ID:
-                        message.append(meta.getId0());
+                        message.append(alarmEntity.getId0());
                         break;
                     case NAME:
-                        message.append(meta.getName());
+                        message.append(alarmEntity.getName());
                 }
             }
         }