+++ oap-server/server-library/library-kubernetes-support/src/main/java/org/apache/skywalking/library/kubernetes/KubernetesEndpoints.java
@@ -39,8 +39,8 @@ public enum KubernetesEndpoints {
     @SneakyThrows
     KubernetesEndpoints() {
         final CacheBuilder<Object, Object> cacheBuilder =
-                CacheBuilder.newBuilder()
-                        .expireAfterAccess(Duration.ofMinutes(3));
+            CacheBuilder.newBuilder()
+                        .expireAfterWrite(Duration.ofMinutes(3));
 
         endpoints = cacheBuilder.build(CacheLoader.from(() -> {
             try (final var kubernetesClient = new KubernetesClientBuilder().build()) {